+++ src/main/java/ru/yandex/practicum/filmorate/dto/NewFilmRequest.java
@@ -11,5 +11,5 @@ public class NewFilmRequest {
     private String description;
     private int duration;
     private LocalDate releaseDate;
-    private int ratingId;
+    private Long ratingId;
 }
